+++ src/main/java/cc/polyfrost/oneconfig/lwjgl/ScissorManager.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public class ScissorManager {
     private static final ArrayList<Scissor> scissors = new ArrayList<>();
 
-    public static Scissor scissor(long vg, int x, int y, int width, int height) {
+    public static Scissor scissor(long vg, float x, float y, float width, float height) {
         Scissor scissor = new Scissor(x, y, width, height);
         if (scissors.contains(scissor)) return scissor;
         scissors.add(scissor);
@@ -28,8 +28,8 @@ public class ScissorManager {
         Scissor finalScissor = new Scissor(scissors.get(0));
         for (int i = 1; i < scissors.size(); i++) {
             Scissor scissor = scissors.get(i);
-            int rightX = Math.min(scissor.x + scissor.width, finalScissor.x + finalScissor.width);
-            int rightY = Math.min(scissor.y + scissor.height, finalScissor.y + finalScissor.height);
+            float rightX = Math.min(scissor.x + scissor.width, finalScissor.x + finalScissor.width);
+            float rightY = Math.min(scissor.y + scissor.height, finalScissor.y + finalScissor.height);
             finalScissor.x = Math.max(finalScissor.x, scissor.x);
             finalScissor.y = Math.max(finalScissor.y, scissor.y);
             finalScissor.width = rightX - finalScissor.x;
